beginning of a sched_wak{eup,ing} -> switch latency analysis
Support for the sched_waking event. It has the same fields as the
sched_wakeup event and the same impact, the only difference between the
two is that sched_waking is guaranteed to be executed in the context of
the waker.

This analysis should display a log, freq distribution and top of all
delays between a sched_wakeup/waking and the corresponding sched_switch.
If sched_waking is available, they will always appear before a
sched_wakeup, and we should ignore successive wakeups of the same
process. That way, we support sched_waking if enabled (>= 4.3),
otherwise we only use sched_wakeup which is a good start for scheduling
latency analysis.

The "cli" part is not done yet, it is a copy of the IRQ module, since it
is the same kind of information we want to present, but needs to be
reworked. Not sure if the "core" part is finished. We should be able to
see the results globally and also the stats per-process (how long it
takes for this process to be woken up avg, min, max, stdev ?).

Let me know, if you need more details.

Thanks !

Julien

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>